Add Seer Station stages preferences during a Run
This changes the scene selection closer to how it's handled in the game
so the blue portal override works smoothly. Also changes the run to not
take into account visiting the Void Fields as counting towards the number
of stages the user wants to simulate. Finally looting the Void Fields and
Bulwark's Ambry are both consolidated in Run._loot_stage.